fix: vue Binding component
