+++ packages/comark-vue/src/plugins/binding.ts
@@ -1,4 +1,4 @@
 export * from 'comark/plugins/binding'
 export { default } from 'comark/plugins/binding'
 
-export { default as Binding } from '../components/Binding.vue'
+export { Binding } from '../components/Binding.ts'
